Extract search input into reusable component with mobile and desktop variants
